# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API documentation for Returns Service App",
        "contact": {
            "email": "admin@adm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/",
            "description": "Returns App API"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "getUserList",
                "description": "Get all Users",
                "operationId": "1df092f5eddb2e29c3c5aa8043dd6d44",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "createUser",
                "description": "Create User",
                "operationId": "3465c05467f56755d8fde6dfe336f784",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "getUserItem",
                "description": "Get User",
                "operationId": "ca775ff26d0144e4b01abbd8c926f7c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "updateUser",
                "description": "Update User",
                "operationId": "acefdff5c10ad10456b5e12506f8ab63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "deleteUser",
                "description": "Delete User",
                "operationId": "b3671c968c691df9358298f3a285c767",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
...
}